Add check for new snooty version
