+++ snooty-cache/src/index.ts
@@ -1,9 +1,11 @@
 // Documentation: https://sdk.netlify.com
+
 import { NetlifyIntegration } from "@netlify/sdk";
-import axios from "axios";
-import { existsSync, readdir } from "fs";
+
+import { readdir } from "fs";
 
 import { promisify } from "util";
+import { checkForNewSnootyVersion } from "./snooty-frontend-version-check";
 
 const readdirAsync = promisify(readdir);
 
@@ -11,32 +13,6 @@ const getCacheFilePaths = (filesPaths: string[]): string[] =>
   filesPaths.filter((filePath) => filePath.endsWith(".cache.gz"));
 
 const integration = new NetlifyIntegration();
-
-interface GitHubCommitResponse {
-  commit: {
-    sha: string;
-  };
-}
-
-async function getLatestSnootyCommit(): Promise<string | undefined> {
-  try {
-    const response = await axios.get<GitHubCommitResponse>(
-      "https://api.github.com/repos/mongodb/snooty/branches/netlify-poc",
-      {
-        headers: {
-          Accept: "application/vnd.github+json",
-          "X-GitHub-Api-Version": "2022-11-28",
-        },
-      }
-    );
-
-    const latestSha = response.data.commit.sha;
-
-    return latestSha;
-  } catch (e) {
-    console.error("Could not retrieve latest SHA", e);
-  }
-}
 
 integration.addBuildEventHandler(
   "onPreBuild",
@@ -57,18 +33,7 @@ integration.addBuildEventHandler(
       cacheFiles.map(async (cacheFile) => await cache.restore(cacheFile))
     );
 
-    console.log("Checking Snooty frontend version");
-    const snootyDirExists = existsSync(`${process.cwd()}/snooty`);
-
-    if (snootyDirExists) {
-      const latestSha = await getLatestSnootyCommit();
-
-      const { stdout: currentSha } = await run.command("git rev-parse HEAD");
-
-      if (currentSha !== latestSha) {
-        await run.command("rm -rf snooty");
-      }
-    }
+    await checkForNewSnootyVersion(run);
   }
 );
